reorganized some unit tests
